# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ecolution API
  description: |
    API documentation for Ecolution, an environmental activity and tracking system built with Next.js 16 (App Router), TypeScript, Prisma, and PostgreSQL.
    
    ### Authentication
    Authentication is handled via **Bearer JWT**.
    - For protected endpoints, pass `Authorization: Bearer <your_jwt_token>` in the request headers.
    - Refresh tokens are managed automatically using HttpOnly cookies (`refresh_token`) set on the `/api/auth` path.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
...
paths:
...
  /api/auth/me:
    get:
      summary: Get current user profile
      description: Retrieve the currently authenticated user's profile details.
      tags:
        - Auth
      security:
        - BearerJWT: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized (Token missing or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '419':
          description: Authentication token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/activity:
    get:
      summary: List activities
      description: Retrieve a list of activities with pagination and filters. Non-admin users can only view their own activities. Admins can view all activities and filter by userId.
      tags:
        - Activity
      security:
        - BearerJWT: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination (starts at 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of activities per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search term matching title or description (case-insensitive)
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          description: Filter by activity category UUID
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by activity status
          required: false
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
        - name: userId
          in: query
          description: Filter by user ID (Only applies to ADMIN roles)
          required: false
          schema:
            type: string
            format: uuid
        - name: sortBy
          in: query
          description: Field to sort by (e.g. createdAt, activityDate)
          required: false
          schema:
            type: string
        - name: sortOrder
          in: query
          description: Sort ordering direction
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of activities retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  meta:
                    type: object
                    required:
                      - page
                      - limit
                      - totalCount
                      - totalPages
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      totalCount:
                        type: integer
                        example: 45
                      totalPages:
                        type: integer
                        example: 5
        '401':
          description: Unauthorized (Token missing or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (insufficient roles or querying another user's activities)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '419':
          description: Authentication token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create activity with image upload
      description: |
        Create a new activity with photo upload using multipart/form-data.
        The service extracts EXIF metadata from the image. If EXIF metadata (GPS coordinates and original date time) is missing, the activity status will be set to **REJECTED**.
      tags:
        - Activity
      security:
        - BearerJWT: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
                - title
                - categoryId
                - activityDate
              properties:
                image:
                  type: string
                  format: binary
                  description: File upload containing the activity image (JPEG, HEIC, TIFF; max 10MB)
                title:
                  type: string
                  minLength: 3
                  maxLength: 150
                  description: Title of the activity
                  example: "Recycling PET Plastic Bottles"
                description:
                  type: string
                  description: Optional description of the activity
                  example: "Collected 20 PET plastic bottles from the neighborhood park."
                categoryId:
                  type: string
                  format: uuid
                  description: UUID of the activity category
                  example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
                activityDate:
                  type: string
                  format: date-time
                  description: Date and time when the activity was performed (ISO 8601)
                  example: "2026-07-04T07:30:00.000Z"
                location:
                  type: string
                  description: Optional textual location description
                  example: "Central Park Block B"
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - id
                      - title
                      - description
                      - categoryId
                      - activityDate
                      - location
                      - status
                      - userId
                      - createdAt
                      - photo
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "b80e81a8-c290-4824-a212-32a829e1fa12"
                      title:
                        type: string
                        example: "Recycling PET Plastic Bottles"
                      description:
                        type: string
                        nullable: true
                        example: "Collected 20 PET plastic bottles from the neighborhood park."
                      categoryId:
                        type: string
                        format: uuid
                        example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
                      activityDate:
                        type: string
                        format: date-time
                        example: "2026-07-04T07:30:00.000Z"
                      location:
                        type: string
                        nullable: true
                        example: "Central Park Block B"
                      status:
                        type: string
                        enum: [PENDING, APPROVED, REJECTED]
                        example: "PENDING"
                      userId:
                        type: string
                        format: uuid
                        example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2026-07-04T12:00:00.000Z"
                      photo:
                        type: object
                        required:
                          - id
                          - imageUrl
                          - latitude
                          - longitude
                          - takenAt
                          - hasExif
                        properties:
                          id:
                            type: string
                            format: uuid
                            example: "c46d3e3a-e275-47e1-8896-1ab3dfa06fb1"
                          imageUrl:
                            type: string
                            example: "/uploads/1720093845-photo.jpg"
                          latitude:
                            type: number
                            nullable: true
                            example: -6.2088
                          longitude:
                            type: number
                            nullable: true
                            example: 106.8456
                          takenAt:
                            type: string
                            format: date-time
                            nullable: true
                            example: "2026-07-04T07:30:00.000Z"
                          hasExif:
                            type: boolean
                            example: true
        '400':
          description: Validation error (e.g. invalid date, size exceeds limit, unsupported format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (Token missing or invalid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Activity category not found or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '419':
          description: Authentication token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/merchant/orders:
    get:
      summary: Get incoming orders for merchant
      description: Retrieve all customer orders containing products from the logged-in merchant's store. UMKM/Admin only.
      tags:
        - Merchant
      responses:
        '200':
          description: Merchant orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/dashboard/admin:
    get:
      summary: Get admin dashboard metrics
      description: Aggregated admin metrics for ecosystem supervision. Admin only.
      tags:
        - Dashboard
      security:
        - BearerJWT: []
      responses:
        '200':
          description: Admin dashboard metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardAdmin'
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT access token in the format `Bearer <token>`.
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - username
        - email
        - role
        - trustScore
        - totalPoint
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
        name:
          type: string
          example: "John Doe"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [USER, UMKM, ADMIN]
          example: "USER"
        trustScore:
          type: integer
          example: 0
        totalPoint:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
          example: "2026-07-04T12:00:00.000Z"
...
    Category:
      type: object
      required:
        - id
        - name
        - pointReward
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        name:
          type: string
          example: "Recycling Plastic"
        description:
          type: string
          nullable: true
          example: "Collect and recycle plastic waste"
        pointReward:
          type: integer
          example: 100
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-07-04T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-04T12:00:00.000Z"
    Activity:
      type: object
      required:
        - id
        - userId
        - categoryId
        - title
        - activityDate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "b80e81a8-c290-4824-a212-32a829e1fa12"
        userId:
          type: string
          format: uuid
          example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
        categoryId:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        title:
          type: string
          example: "Morning Clean Up"
        description:
          type: string
          nullable: true
          example: "Clean up activities in local park"
        activityDate:
          type: string
          format: date-time
          example: "2026-07-04T07:00:00.000Z"
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
          example: "PENDING"
        adminNote:
          type: string
          nullable: true
          example: null
        location:
          type: string
          nullable: true
          example: "Grand Park Jakarta"
        createdAt:
          type: string
          format: date-time
          example: "2026-07-04T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-04T12:00:00.000Z"
        category:
          $ref: '#/components/schemas/Category'
        photos:
          type: array
          items:
            type: object
            required:
              - id
              - activityId
              - imageUrl
              - hasExif
              - uploadedAt
            properties:
              id:
                type: string
                format: uuid
                example: "c46d3e3a-e275-47e1-8896-1ab3dfa06fb1"
              activityId:
                type: string
                format: uuid
                example: "b80e81a8-c290-4824-a212-32a829e1fa12"
              imageUrl:
                type: string
                example: "/uploads/1720093845-photo.jpg"
              latitude:
                type: number
                nullable: true
                example: -6.2088
              longitude:
                type: number
                nullable: true
                example: 106.8456
              takenAt:
                type: string
                format: date-time
                nullable: true
                example: "2026-07-04T11:45:00.000Z"
              hasExif:
                type: boolean
                example: true
              uploadedAt:
                type: string
                format: date-time
                example: "2026-07-04T12:00:00.000Z"
        verification:
          type: object
          nullable: true
          required:
            - id
            - activityId
            - verifiedBy
            - method
            - verifiedAt
          properties:
            id:
              type: string
              format: uuid
              example: "f290f1ee-6c54-4b01-90e6-d701748f0853"
            activityId:
              type: string
              format: uuid
              example: "b80e81a8-c290-4824-a212-32a829e1fa12"
            verifiedBy:
              type: string
              format: uuid
              example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
            method:
              type: string
              enum: [MANUAL, AI]
              example: "AI"
            confidence:
              type: number
              nullable: true
              example: 88.5
            note:
              type: string
              nullable: true
              example: "Auto verified by AI"
            verifiedAt:
              type: string
              format: date-time
              example: "2026-07-04T12:05:00.000Z"
        pointHistory:
          type: object
          nullable: true
          required:
            - id
            - userId
            - type
            - point
            - createdAt
          properties:
            id:
              type: string
              format: uuid
              example: "e290f1ee-6c54-4b01-90e6-d701748f0854"
            userId:
              type: string
              format: uuid
              example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
            activityId:
              type: string
              format: uuid
              nullable: true
              example: "b80e81a8-c290-4824-a212-32a829e1fa12"
            type:
              type: string
              enum: [EARN, REDEEM, BONUS, PENALTY]
              example: "EARN"
            point:
              type: integer
              example: 100
            description:
              type: string
              nullable: true
              example: "Point reward for activity: Recycling Plastic"
            createdAt:
              type: string
              format: date-time
              example: "2026-07-04T12:05:00.000Z"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "UNAUTHORIZED"
            message:
              type: string
              example: "Token missing or invalid"
            details:
              type: object
              nullable: true
              additionalProperties: true
...
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - merchantId
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "p80b82f9-3dc8-4f24-91b7-d1587d558b8b"
        name:
          type: string
          example: "Tumbler Stainless Ramah Lingkungan"
        description:
          type: string
          nullable: true
          example: "Tumbler stainless steel tahan panas dan dingin 500ml."
        price:
          type: integer
          example: 85000
        stock:
          type: integer
          example: 50
        merchantId:
          type: string
          format: uuid
          example: "e80b82f9-3dc8-4f24-91b7-d1587d558b8a"
        imageThumbnail:
          type: string
          nullable: true
          example: "/products/tumbler-thumb.jpg"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        merchant:
          type: object
          properties:
            businessName:
              type: string
              example: "Eco Store Indonesia"
...
    Order:
      type: object
      required:
        - id
        - orderNumber
        - userId
        - status
        - totalPrice
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "o80b82f9-3dc8-4f24-91b7-d1587d558b8c"
        orderNumber:
          type: string
          example: "ORD-20260705-0001"
        userId:
          type: string
          format: uuid
          example: "a80b82f9-3dc8-4f24-91b7-d1587d558b8f"
        status:
          type: string
          enum: [PENDING, PAID, PROCESSING, COMPLETED, CANCELLED]
          example: "PENDING"
        totalPrice:
          type: integer
          example: 85000
        note:
          type: string
          nullable: true
          example: "Alamat: Jl. Sudirman No 10 | Catatan: Ketuk pintu sebelum masuk"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
              price:
                type: integer
              product:
                $ref: '#/components/schemas/Product'
...
    DashboardAdmin:
      type: object
      properties:
        totalUsers:
          type: integer
          example: 150
        totalMerchants:
          type: integer
          example: 12
        totalPointsCirculation:
          type: integer
          example: 25000
        activitiesCount:
          type: object
          properties:
            PENDING:
              type: integer
              example: 5
            APPROVED:
              type: integer
              example: 45
            REJECTED:
              type: integer
              example: 2
            TOTAL:
              type: integer
              example: 52
        topUsers:
          type: array
          items:
            type: object
...